rainy: Research Compendium for the Paper 'The relationship between rain intensity and Ascochyta blight in chickpea development'

Documents the workflow and processes involved in the analysis of data for the paper '', <https://zenodo.org/badge/latestdoi/378052232>.

Author

Maintainer: Ihsanul Khaliq ihsan.khaliq@usq.edu.au (ORCID)

Authors:

Kevin Moore kevin.moore@dpi.nsw.gov.au (ORCID)

Adam H. Sparks adam.sparks@usq.edu.au (ORCID) [conceptor]

Other contributors:

Grains Research and Development Corporation (Provided support through funding GRDC Project USQ1903-003RTX) [funder, copyright holder]

Western Australia Agriculture Authority (Provided support through Adam Sparks' time.) [funder, copyright holder]

University of Southern Queensland [copyright holder]